requirement.txt
streamlit
pandas
boto3
plotly
Pillow
